pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.2.1"
description = ""
authors = ["roaddust2 <roaddust2@yahoo.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
pytest-lazy-fixture = "^0.6.3"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
